+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Changed
 - [package] Updated dependencies.
 - [package][data] `twitter` proprty is no longer a part of the `package.json`. It's moved into data `SOCIAL.twitter.handle` instead.
+- [package] Remove `grunt-contrib-uglify` in favor of `uglify-js` bundled with `jspm`, rel [#294](https://github.com/LotusTM/Kotsu/issues/294)
 - [styles] Default `.Icon--left` and `.Icon--right` classes now have more reasonable margins.
 - [scripts] ES6 in, CoffeeScript out.